# optimize utf8_length_from_latin1 for short strings