@lemire 2024-08-13
cc @ronag